Move instances from data. Add emptyC, mapFilterC.
